Use encoding value from decoding "git diff" output
We used to determine the correct encoding by reading and decoding the
actual file contents. This is probably an unnecessary step since we
can just use the actual encoding we used for decoding the "git diff"
output. (We also could have just used the encoding Sublime uses for the
view by reading `view.encoding()` but then we need a fallback mechanism
if no view into that file is currently open.)

We also refactor the underlying `decode_stdout` method which is used
for every git command. 1. Instead of "latin-1" (Western Europe) we use
`getpreferredencoding` from `locale`. "latin-1" was always a totally
arbitrary choice anyway. 2. We remove the `silent_fallback` setting
which is annoying. If you have non-utf8 files you always need to flip
that setting to `True`, and if you don't you never notice.