Optimize toggle behavior
Let `cached` have three states, `None, True, False`.  We expect that
the common key-combo calls the action without any arguments, t.i. with
`None`.

Calling with `None` should always toggle the current active view *if*
it is an inline view, otherwise it shows unstaged changes. (This mode
is most useful for this kind of view.)

Explicit values should switch to the desired mode *if* in the other
mode, otherwise close the view.